+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## WIP
 
 - Refactoring
+- FIXED
+  - Add **icupy.icu.MeasureUnit.create_milligram_ofglucose_per_deciliter()**
+  - Add **icupy.icu.MeasureUnit.get_milligram_ofglucose_per_deciliter()**
 
 ## v0.16.0 / 2023-04-18